pyproject.toml
[project]
name = "video-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "ty>=0.0.17",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]
